+++ bot/whoisondutytodayslackbot.rb
@@ -102,6 +102,10 @@ class WhoIsOnDutyTodaySlackBot < SlackRubyBot::Bot
     WhoIsOnDutyTodaySlackBotModule::Commands::ChannelLabelsList.call(client: client, data: data, match: match)
   end
 
+  command "channel labels merge" do |client, data, match|
+    WhoIsOnDutyTodaySlackBotModule::Commands::ChannelLabelsMerge.call(client: client, data: data, match: match)
+  end
+
   command(/.*/) do |client, data|
     WhoIsOnDutyTodaySlackBotModule::Commands::Unknown.call(client: client, data: data)
   end